Move some things around.
